refactor(di): move code to suitable place
