+++ packages/di/src/container/container.class.ts
@@ -1,6 +1,12 @@
-import { GenericProvider, Token, ExportableCnf, ImportableCnf } from '../types';
+import {
+  GenericProvider,
+  Token,
+  ExportableCnf,
+  ClassifiedProvider,
+  ImportableCnf,
+  ClassType,
+} from '../types';
 import { DiResolvers } from '../consts/di-resolver.const';
-import { ContainerBag } from './props/container-bag.prop';
 import { addExtModule } from './funcs/addExtModule.func';
 import { addModuleToMap } from './funcs/addModuleToMap.func';
 import { addModule } from './funcs/addModule.func';
@@ -17,7 +23,34 @@ import { resolveUseFuncProvider } from './funcs/resolveUseFuncProvider.func';
 import { resolveUseValueProvider } from './funcs/resolveUseValueProvider.func';
 import { resolveUseExistingProvider } from './funcs/resolveUseExistingProvider.func';
 
-export class Container extends ContainerBag {
+export class Container {
+  /**
+   * A store for cached resolved values.
+   */
+  protected _resolvedValues = new Map<Token, any>();
+
+  /**
+   * List of providers containing information for how to create dependency value.
+   */
+  protected _providers = new Map<Token, ClassifiedProvider>();
+
+  /**
+   * List of modules that current module/container use as reference for resolving
+   * dependency value.
+   */
+  protected _extModules = new Map<ClassType, Container>();
+
+  /**
+   * `_parentModule` is a module that this module extend from.
+   */
+  protected _parentModule: Container;
+
+  constructor(private moduleClass?: ClassType) {}
+
+  public getModuleClass() {
+    return this.moduleClass;
+  }
+
   /**
    * Convenient method for adding multiple providers into this container.
    */
