+++ packages/di/src/container/index.ts
@@ -1,2 +1,1 @@
 export { Container } from './container.class';
-export { clearModuleMap } from './props/module-map.static';
