+++ packages/di/src/internal.ts
@@ -4,3 +4,4 @@ export * from './consts/error.const';
 export * from './forward-ref';
 export * from './utils';
 export * from './types';
+export * from './module-map';
